Added "Доходы"("Income") tab functionality
1. income items table
1.1. released common actions: add, remove, update income data
2. spend filters table
2.1. after column (filter type) removal all filters are assigned to
others type ("Другое")

+++ Balansiq/Balansiq/DB/Entities/IItem.cs
@@ -25,7 +25,7 @@ namespace Balansiq.DB.Entities
         [Ignore]
         public static KeyValuePair<Balansiq.DB.DataTableType, string>? ForeignKey { get { return null; } }
 
-        public IItem(int? id)
+        public IItem(long? id)
         {
             Id = id;
         }
